refactor: ApproveStatus 패키지 artist/users 하위로 이동
