+++ src/main/java/com/prography/lighton/artist/admin/application/ArtistApplicationQueryUseCase.java
@@ -1,8 +1,8 @@
 package com.prography.lighton.artist.admin.application;
 
-import com.prography.lighton.artist.admin.domain.enums.ApproveStatus;
 import com.prography.lighton.artist.admin.presentation.GetArtistApplicationDetailResponseDTO;
 import com.prography.lighton.artist.admin.presentation.GetArtistApplicationListResponseDTO;
+import com.prography.lighton.artist.users.domain.entity.enums.ApproveStatus;
 
 public interface ArtistApplicationQueryUseCase {
